requirements.txt
ufal.udpipe == 1.2.0.3
